Added Delete and DeleteAll buttons classes and refactored folders
